requirements.txt
biopython==1.79
numpy==1.23.0
Pillow==9.1.1
reportlab==3.6.10
